requirements.txt
python-telegram-bot==21.6
requests==2.32.3
feedparser==6.0.11
APScheduler==3.10.4
SQLAlchemy==2.0.36
python-dotenv==1.0.1

